+++ src/internal/localsession/session_state.go
@@ -2,53 +2,18 @@ package localsession
 
 import (
 	"errors"
-	"sync"
 
 	"github.com/rinq/rinq-go/src/internal/attributes"
 	"github.com/rinq/rinq-go/src/rinq"
 	"github.com/rinq/rinq-go/src/rinq/ident"
 )
 
-// State represents a session's revisioned state.
-type State struct {
-	mutex     sync.RWMutex
-	ref       ident.Ref
-	attrs     attributes.Catalog
-	seq       uint32
-	destroyed chan struct{}
-	logger    rinq.Logger
-}
-
-// Ref returns the most recent session-ref.
-// The ref's revision increments each time a call to TryUpdate() succeeds.
-func (s *State) Ref() ident.Ref {
-	s.mutex.RLock()
-	defer s.mutex.RUnlock()
-
-	return s.ref
-}
-
-// NextMessageID generates a unique message ID from the current session-ref.
-func (s *State) NextMessageID() (ident.MessageID, attributes.Catalog) {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-
-	s.seq++
-	return s.ref.Message(s.seq), s.attrs
-}
-
-// Head returns the most recent revision.
-// It is conceptually equivalent to s.At(s.Ref().Rev).
-func (s *State) Head() rinq.Revision {
-	s.mutex.RLock()
-	defer s.mutex.RUnlock()
-
-	return &revision{s.ref, s, s.attrs, s.logger}
-}
+// This file contains methods of the Session struct that are not defined in
+// rinq.Session. See session.go for the implementation of rinq.Session.
 
 // At returns a revision representing the state at a specific revision
 // number. The revision can not be newer than the current session-ref.
-func (s *State) At(rev ident.Revision) (rinq.Revision, error) {
+func (s *Session) At(rev ident.Revision) (rinq.Revision, error) {
 	s.mutex.RLock()
 	defer s.mutex.RUnlock()
 
@@ -65,7 +30,7 @@ func (s *State) At(rev ident.Revision) (rinq.Revision, error) {
 }
 
 // Attrs returns all attributes at the most recent revision.
-func (s *State) Attrs() (ident.Ref, attributes.Catalog) {
+func (s *Session) Attrs() (ident.Ref, attributes.Catalog) {
 	s.mutex.RLock()
 	defer s.mutex.RUnlock()
 
@@ -73,7 +38,7 @@ func (s *State) Attrs() (ident.Ref, attributes.Catalog) {
 }
 
 // AttrsIn returns all attributes in the ns namespace at the most recent revision.
-func (s *State) AttrsIn(ns string) (ident.Ref, attributes.VTable) {
+func (s *Session) AttrsIn(ns string) (ident.Ref, attributes.VTable) {
 	s.mutex.RLock()
 	defer s.mutex.RUnlock()
 
@@ -85,25 +50,19 @@ func (s *State) AttrsIn(ns string) (ident.Ref, attributes.VTable) {
 //
 // The operation fails if ref is not the current session-ref, attrs includes
 // changes to frozen attributes, or the session has been destroyed.
-func (s *State) TryUpdate(
-	ref ident.Ref,
-	ns string,
-	attrs attributes.List,
-) (rinq.Revision, *attributes.Diff, error) {
+func (s *Session) TryUpdate(rev ident.Revision, ns string, attrs attributes.List) (rinq.Revision, *attributes.Diff, error) {
 	s.mutex.Lock()
 	defer s.mutex.Unlock()
 
-	select {
-	case <-s.destroyed:
+	if s.isDestroyed {
 		return nil, nil, rinq.NotFoundError{ID: s.ref.ID}
-	default:
 	}
 
-	if ref != s.ref {
-		return nil, nil, rinq.StaleUpdateError{Ref: ref}
+	if rev != s.ref.Rev {
+		return nil, nil, rinq.StaleUpdateError{Ref: s.ref.ID.At(rev)}
 	}
 
-	nextRev := ref.Rev + 1
+	nextRev := rev + 1
 	nextAttrs := s.attrs[ns].Clone()
 	diff := attributes.NewDiff(ns, nextRev)
 
@@ -115,7 +74,7 @@ func (s *State) TryUpdate(
 		}
 
 		if entry.IsFrozen {
-			return nil, nil, rinq.FrozenAttributesError{Ref: ref}
+			return nil, nil, rinq.FrozenAttributesError{Ref: s.ref.ID.At(rev)}
 		}
 
 		entry.Attr = attr
@@ -129,7 +88,7 @@ func (s *State) TryUpdate(
 	}
 
 	s.ref.Rev = nextRev
-	s.seq = 0
+	s.msgSeq = 0
 
 	if !diff.IsEmpty() {
 		s.attrs = s.attrs.WithNamespace(ns, nextAttrs)
@@ -148,32 +107,27 @@ func (s *State) TryUpdate(
 //
 // The operation fails if ref is not the current session-ref, there are any
 // frozen attributes, or the session has been destroyed.
-func (s *State) TryClear(
-	ref ident.Ref,
-	ns string,
-) (rinq.Revision, *attributes.Diff, error) {
+func (s *Session) TryClear(rev ident.Revision, ns string) (rinq.Revision, *attributes.Diff, error) {
 	s.mutex.Lock()
 	defer s.mutex.Unlock()
 
-	select {
-	case <-s.destroyed:
+	if s.isDestroyed {
 		return nil, nil, rinq.NotFoundError{ID: s.ref.ID}
-	default:
 	}
 
-	if ref != s.ref {
-		return nil, nil, rinq.StaleUpdateError{Ref: ref}
+	if rev != s.ref.Rev {
+		return nil, nil, rinq.StaleUpdateError{Ref: s.ref.ID.At(rev)}
 	}
 
 	attrs := s.attrs[ns]
-	nextRev := ref.Rev + 1
+	nextRev := rev + 1
 	nextAttrs := attributes.VTable{}
 	diff := attributes.NewDiff(ns, nextRev)
 
 	for _, entry := range attrs {
 		if entry.Value != "" {
 			if entry.IsFrozen {
-				return nil, nil, rinq.FrozenAttributesError{Ref: ref}
+				return nil, nil, rinq.FrozenAttributesError{Ref: s.ref.ID.At(rev)}
 			}
 
 			entry.Value = ""
@@ -185,7 +139,7 @@ func (s *State) TryClear(
 	}
 
 	s.ref.Rev = nextRev
-	s.seq = 0
+	s.msgSeq = 0
 
 	if !diff.IsEmpty() {
 		s.attrs = s.attrs.WithNamespace(ns, nextAttrs)
@@ -203,39 +157,43 @@ func (s *State) TryClear(
 //
 // The operation fails if ref is not the current session-ref. It is not an
 // error to destroy an already-destroyed session.
-func (s *State) TryDestroy(ref ident.Ref) error {
+//
+// first is true if this call caused the session to be destroyed.
+func (s *Session) TryDestroy(rev ident.Revision) (bool, error) {
 	s.mutex.Lock()
 	defer s.mutex.Unlock()
 
-	if ref != s.ref {
-		return rinq.StaleUpdateError{Ref: ref}
+	if rev != s.ref.Rev {
+		return false, rinq.StaleUpdateError{Ref: s.ref.ID.At(rev)}
 	}
 
-	select {
-	case <-s.destroyed:
-	default:
-		close(s.destroyed)
+	if s.isDestroyed {
+		return false, nil
 	}
 
-	return nil
+	s.destroy()
+
+	return true, nil
 }
 
-// ForceDestroy forcefully destroys the session, preventing further updates.
-// It is not an error to destroy an already-destroyed session.
-func (s *State) ForceDestroy() bool {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
+// destroy marks the session as destroyed removes any callbacks registered with
+// the command and notification subsystems.
+func (s *Session) destroy() {
+	s.isDestroyed = true
 
-	select {
-	case <-s.destroyed:
-		return false
-	default:
-		close(s.destroyed)
-		return true
-	}
+	s.invoker.SetAsyncHandler(s.ref.ID, nil)
+	_ = s.listener.UnlistenAll(s.ref.ID)
+
+	go func() {
+		// close the done channel only after all pending calls have finished
+		s.calls.Wait()
+		close(s.done)
+	}()
 }
 
-// Destroyed returns a channel that is closed when the session is destroyed.
-func (s *State) Destroyed() <-chan struct{} {
-	return s.destroyed
+// nextMessageID returns a new unique message ID generated from the current
+// session-ref, and the attributes as they existed at that ref.
+func (s *Session) nextMessageID() ident.MessageID {
+	s.msgSeq++
+	return s.ref.Message(s.msgSeq)
 }