+++ src/internal/localsession/store.go
@@ -3,74 +3,68 @@ package localsession
 import (
 	"sync"
 
-	revisionpkg "github.com/rinq/rinq-go/src/internal/revision"
+	"github.com/rinq/rinq-go/src/internal/revisions"
 	"github.com/rinq/rinq-go/src/rinq"
 	"github.com/rinq/rinq-go/src/rinq/ident"
 )
 
-// Store is a collection of sessions and their state.
-type Store interface {
-	revisionpkg.Store
-
-	Add(rinq.Session, *State)
-	Remove(ident.SessionID)
-	Get(ident.SessionID) (rinq.Session, *State, bool)
-	Each(fn func(rinq.Session, *State))
-}
-
-type store struct {
-	mutex   sync.RWMutex
-	entries map[ident.SessionID]storeEntry
+// Store is a collection of local sessions which provides an implementation
+// of revisions.Store.
+type Store struct {
+	mutex    sync.RWMutex
+	sessions map[ident.SessionID]*Session
 }
 
 // NewStore returns a new session store.
-func NewStore() Store {
-	return &store{
-		entries: map[ident.SessionID]storeEntry{},
+func NewStore() *Store {
+	return &Store{
+		sessions: map[ident.SessionID]*Session{},
 	}
 }
 
-type storeEntry struct {
-	Session rinq.Session
-	State   *State
-}
-
-func (s *store) Add(sess rinq.Session, state *State) {
+// Add adds a session to the store.
+func (s *Store) Add(sess *Session) {
 	s.mutex.Lock()
 	defer s.mutex.Unlock()
 
-	s.entries[sess.ID()] = storeEntry{sess, state}
+	s.sessions[sess.ID()] = sess
 }
 
-func (s *store) Remove(id ident.SessionID) {
+// Remove removes a session to from the store.
+func (s *Store) Remove(id ident.SessionID) {
 	s.mutex.Lock()
 	defer s.mutex.Unlock()
 
-	delete(s.entries, id)
+	delete(s.sessions, id)
 }
 
-func (s *store) Get(id ident.SessionID) (rinq.Session, *State, bool) {
+// Get fetches a session from the store by its ID.
+func (s *Store) Get(id ident.SessionID) (sess *Session, ok bool) {
 	s.mutex.RLock()
 	defer s.mutex.RUnlock()
 
-	e, ok := s.entries[id]
-	return e.Session, e.State, ok
+	sess, ok = s.sessions[id]
+	return
 }
 
-func (s *store) Each(fn func(rinq.Session, *State)) {
+// Each calls fn(sess) for each session in the store.
+func (s *Store) Each(fn func(*Session)) {
 	s.mutex.RLock()
 	defer s.mutex.RUnlock()
 
-	for _, e := range s.entries {
-		fn(e.Session, e.State)
+	for _, sess := range s.sessions {
+		fn(sess)
 	}
 }
 
-func (s *store) GetRevision(ref ident.Ref) (rinq.Revision, error) {
-	_, state, ok := s.Get(ref.ID)
-	if ok {
-		return state.At(ref.Rev)
+// GetRevision returns the session revision for the given ref.
+func (s *Store) GetRevision(ref ident.Ref) (rinq.Revision, error) {
+	s.mutex.RLock()
+	defer s.mutex.RUnlock()
+
+	if sess, ok := s.sessions[ref.ID]; ok {
+		return sess.At(ref.Rev)
 	}
 
-	return revisionpkg.Closed(ref), nil
+	return revisions.Closed(ref.ID), nil
 }